Attempt to fix power consumption.
I determined my moisture sensors were drawing too much power, as the batteries
should be lasting many weeks but were going dead within a day. This update
was an attempt to remedy the situation by controlling the power to the
sensors through a GPIO pin instead of the 3v3, turning on and off their
power as needed. I furthermore manually turned on and off wifi capabilities
in case the deepsleep mode was not doing this automatically. Finnaly I
added a line to ensure no additional wakeup sources were unintentionally set.

I will monitor the battery life to see if there are any improvements.

- Now that the sesnors are drawing power through a GPIO pin, I fear that
their moisture readings may be changed. I will also be monitoring this
to see if that is the case.